--- a/tools/src/main/scala/scala/scalanative/interflow/Inline.scala
+++ b/tools/src/main/scala/scala/scalanative/interflow/Inline.scala
@@ -20,17 +20,17 @@ trait Inline { self: Interflow =>
       .fold[Boolean] {
         false
       } { defn =>
-        val isCtor = originalName(name) match {
+        lazy val isCtor = originalName(name) match {
           case Global.Member(_, sig) if sig.isCtor || sig.isImplCtor =>
             true
           case _ =>
             false
         }
-        val isSmall =
+        lazy val isSmall =
           defn.insts.size <= 8
         val isExtern =
           defn.attrs.isExtern
-        val hasVirtualArgs =
+        lazy val hasVirtualArgs =
           args.exists(_.isInstanceOf[Val.Virtual])
         val noOpt =
           defn.attrs.opt == Attr.NoOpt
@@ -40,15 +40,15 @@ trait Inline { self: Interflow =>
           defn.attrs.inlineHint == Attr.AlwaysInline
         val hintInline =
           defn.attrs.inlineHint == Attr.InlineHint
-        val isRecursive =
+        lazy val isRecursive =
           hasContext(s"inlining ${name.show}")
-        val isBlacklisted =
+        lazy val isBlacklisted =
           this.isBlacklisted(name)
-        val calleeTooBig =
+        lazy val calleeTooBig =
           defn.insts.size > 8192
-        val callerTooBig =
+        lazy val callerTooBig =
           mergeProcessor.currentSize() > 8192
-        val hasUnwind = defn.insts.exists {
+        lazy val hasUnwind = defn.insts.exists {
           case Inst.Let(_, _, unwind)   => unwind ne Next.None
           case Inst.Throw(_, unwind)    => unwind ne Next.None
           case Inst.Unreachable(unwind) => unwind ne Next.None
@@ -57,37 +57,38 @@ trait Inline { self: Interflow =>
 
         val shall = mode match {
           case build.Mode.Debug =>
-            isCtor || alwaysInline
+            alwaysInline || isCtor
           case build.Mode.ReleaseFast =>
-            isCtor || alwaysInline || hintInline || isSmall
+            alwaysInline || hintInline || isSmall || isCtor
           case build.Mode.ReleaseFull =>
-            isCtor || alwaysInline || hintInline || isSmall || hasVirtualArgs
+            alwaysInline || hintInline || isSmall || isCtor || hasVirtualArgs
         }
-        val shallNot =
+        lazy val shallNot =
           noOpt || noInline || isRecursive || isBlacklisted || calleeTooBig || callerTooBig || isExtern || hasUnwind
-
-        if (shall) {
-          if (shallNot) {
-            log(s"not inlining ${name.show}, because:")
-            if (noInline) {
-              log("* has noinline attr")
+        withLogger { logger =>
+          if (shall) {
+            if (shallNot) {
+              logger(s"not inlining ${name.show}, because:")
+              if (noInline) {
+                logger("* has noinline attr")
+              }
+              if (isRecursive) {
+                logger("* is recursive")
+              }
+              if (isBlacklisted) {
+                logger("* is blacklisted")
+              }
+              if (callerTooBig) {
+                logger("* caller is too big")
+              }
+              if (calleeTooBig) {
+                logger("* callee is too big")
+              }
             }
-            if (isRecursive) {
-              log("* is recursive")
-            }
-            if (isBlacklisted) {
-              log("* is blacklisted")
-            }
-            if (callerTooBig) {
-              log("* caller is too big")
-            }
-            if (calleeTooBig) {
-              log("* callee is too big")
-            }
+          } else {
+            logger(
+              s"no reason to inline ${name.show}(${args.map(_.show).mkString(",")})")
           }
-        } else {
-          log(
-            s"no reason to inline ${name.show}(${args.map(_.show).mkString(",")})")
         }
 
         shall && !shallNot