--- a/tools/src/main/scala/scala/scalanative/interflow/State.scala
+++ b/tools/src/main/scala/scala/scalanative/interflow/State.scala
@@ -9,6 +9,8 @@ import scalanative.codegen.Lower
 
 final class State(block: Local) {
   var fresh   = Fresh(block.id)
+  /* Performance Note: OpenHashMap/LongMap/AnyRefMap have a faster clone()
+   * operation. This really makes a difference on fullClone() */
   var heap    = mutable.LongMap.empty[Instance]
   var locals  = mutable.OpenHashMap.empty[Local, Val]
   var delayed = mutable.AnyRefMap.empty[Op, Val]