Adds some performance optimizations to Interflow, mainly regarding hash map cloning and avoiding unnecessary operations when logging is disabled.

I ran this on a [small codebase](https://github.com/JD557/minart/blob/06ecc9edbe99ffaa221f6e239895085e81bf71fa/examples/colorsquare/shared/src/main/scala/eu/joaocosta/minart/examples/ColorSquare.scala) and got some nice performance improvements on `release-full`.

|  | Optimization (Avg) | Total (Avg) |
|-|-|-|
| 0.4.0 | 170934ms | 226639ms |
| This Branch | 90806ms | 156839ms |

It might be nice to also try this on a large codebase. Any ideas?

-----

Full times of the 3 runs (SBT was not restarted between runs to keep the JIT warm, it was restarted between tests):

# 0.4.0

## Run 1

```
[info] Linking (2373 ms)
[info] Discovered 1083 classes and 9147 methods
[info] Optimizing (release-full mode) (173719 ms)
[info] Generating intermediate code (7042 ms)
[info] Produced 8 files
[info] Compiling to native code (15208 ms)
[info] Linking native code (immix gc, thin lto) (33952 ms)
[info] Total (232533 ms)
```

## Run 2

```
[info] Linking (1116 ms)
[info] Discovered 1083 classes and 9147 methods
[info] Optimizing (release-full mode) (169880 ms)
[info] Generating intermediate code (5002 ms)
[info] Produced 8 files
[info] Compiling to native code (13307 ms)
[info] Linking native code (immix gc, thin lto) (34893 ms)
[info] Total (224436 ms)
```

## Run 3

```
[info] Linking (1167 ms)
[info] Discovered 1083 classes and 9147 methods
[info] Optimizing (release-full mode) (169202 ms)
[info] Generating intermediate code (4851 ms)
[info] Produced 8 files
[info] Compiling to native code (13107 ms)
[info] Linking native code (immix gc, thin lto) (34634 ms)
[info] Total (223112 ms)
```

# New Branch

## Run 1

```
[info] Linking (2868 ms)
[info] Discovered 1083 classes and 9147 methods
[info] Optimizing (release-full mode) (94776 ms)
[info] Generating intermediate code (8554 ms)
[info] Produced 8 files
[info] Compiling to native code (13660 ms)
[info] Linking native code (immix gc, thin lto) (45077 ms)
[info] Total (165277 ms)
```

## Run 2

```
[info] Linking (1186 ms)
[info] Discovered 1083 classes and 9147 methods
[info] Optimizing (release-full mode) (89203 ms)
[info] Generating intermediate code (5405 ms)
[info] Produced 8 files
[info] Compiling to native code (13693 ms)
[info] Linking native code (immix gc, thin lto) (45296 ms)
[info] Total (154986 ms)
```

## Run 3

```
[info] Linking (971 ms)
[info] Discovered 1083 classes and 9147 methods
[info] Optimizing (release-full mode) (88440 ms)
[info] Generating intermediate code (4288 ms)
[info] Produced 8 files
[info] Compiling to native code (12454 ms)
[info] Linking native code (immix gc, thin lto) (44035 ms)
[info] Total (150305 ms)
```