@ekrich 2021-01-28
Thank you for the PR. It looks like you need to run `scalafmt` on the code. Try `$ scripts/scalafmt` inside the `scala-native` directory.