@JD557 2021-01-29
>Thank you for the PR. It looks like you need to run scalafmt on the code. Try $ scripts/scalafmt inside the scala-native directory.

>As @ekrich already suggested you'll need to run scipts/scalafmt from the project directory and push fixed formatting.

Done in 73f3f53.

> Also since it's your first PR to Scala Native, I'd like to welcome on board!

Thank you :)